Added Unit-Tests

+++ backend/microservices/user-meeting/src/main/java/de/share_your_idea/user_meeting/repository/UserEntityRepository.java
@@ -19,5 +19,5 @@ public interface UserEntityRepository extends JpaRepository<UserEntity, Long> {
 
     @Modifying
     @Transactional(readOnly = false)
-    Long deleteUserEntityByUsername(String username);
+    int deleteUserEntityByUsername(String username);
 }
